requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --all-extras pyproject.toml
#
accessible-pygments==0.0.5
    # via pydata-sphinx-theme
alabaster==0.7.16
    # via sphinx
argcomplete==3.3.0
    # via commitizen
babel==2.15.0
    # via
    #   pydata-sphinx-theme
    #   sphinx
beautifulsoup4==4.12.3
    # via pydata-sphinx-theme
black==24.4.2
    # via rocm-docs-core (pyproject.toml)
breathe==4.35.0
    # via rocm-docs-core (pyproject.toml)
build==1.2.1
    # via
    #   pip-tools
    #   rocm-docs-core (pyproject.toml)
certifi==2024.7.4
    # via requests
cffi==1.16.0
    # via
    #   cryptography
    #   pynacl
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.3.2
    # via
    #   commitizen
    #   requests
click==8.1.7
    # via
    #   black
    #   click-log
    #   doxysphinx
    #   pip-tools
    #   sphinx-external-toc
click-log==0.4.0
    # via doxysphinx
colorama==0.4.6
    # via commitizen
commitizen==3.28.0
    # via rocm-docs-core (pyproject.toml)
cryptography==43.0.0
    # via pyjwt
decli==0.6.2
    # via commitizen
deprecated==1.2.14
    # via pygithub
distlib==0.3.8
    # via virtualenv
docutils==0.21.2
    # via
    #   breathe
    #   myst-parser
    #   pydata-sphinx-theme
    #   sphinx
doxysphinx==3.3.10
    # via rocm-docs-core (pyproject.toml)
exceptiongroup==1.2.2
    # via pytest
fastjsonschema==2.20.0
    # via rocm-docs-core (pyproject.toml)
filelock==3.15.4
    # via virtualenv
gitdb==4.0.11
    # via gitpython
gitpython==3.1.43
    # via rocm-docs-core (pyproject.toml)
identify==2.6.0
    # via pre-commit
idna==3.7
    # via requests
imagesize==1.4.1
    # via sphinx
importlib-metadata==8.2.0
    # via commitizen
iniconfig==2.0.0
    # via pytest
isort==5.13.2
    # via rocm-docs-core (pyproject.toml)
jinja2==3.1.4
    # via
    #   commitizen
    #   myst-parser
    #   sphinx
libsass==0.22.0
    # via doxysphinx
lxml==4.9.4
    # via doxysphinx
markdown-it-py==3.0.0
    # via
    #   mdit-py-plugins
    #   myst-parser
markupsafe==2.1.5
    # via jinja2
mdit-py-plugins==0.4.1
    # via myst-parser
mdurl==0.1.2
    # via markdown-it-py
mpire==2.10.2
    # via doxysphinx
mypy==1.11.1
    # via rocm-docs-core (pyproject.toml)
mypy-extensions==1.0.0
    # via
    #   black
    #   mypy
myst-parser==3.0.1
    # via rocm-docs-core (pyproject.toml)
nodeenv==1.9.1
    # via pre-commit
numpy==1.26.4
    # via doxysphinx
packaging==24.1
    # via
    #   black
    #   build
    #   commitizen
    #   pydata-sphinx-theme
    #   pytest
    #   sphinx
pathspec==0.12.1
    # via black
pip-tools==7.4.1
    # via rocm-docs-core (pyproject.toml)
platformdirs==4.2.2
    # via
    #   black
    #   virtualenv
pluggy==1.5.0
    # via pytest
pre-commit==3.8.0
    # via rocm-docs-core (pyproject.toml)
prompt-toolkit==3.0.36
    # via questionary
pycparser==2.22
    # via cffi
pydata-sphinx-theme==0.15.4
    # via
    #   rocm-docs-core (pyproject.toml)
    #   sphinx-book-theme
pygithub==2.3.0
    # via rocm-docs-core (pyproject.toml)
pygments==2.18.0
    # via
    #   accessible-pygments
    #   mpire
    #   pydata-sphinx-theme
    #   sphinx
pyjson5==1.6.6
    # via doxysphinx
pyjwt[crypto]==2.9.0
    # via pygithub
pynacl==1.5.0
    # via pygithub
pyparsing==3.1.2
    # via
    #   doxysphinx
    #   sphinxcontrib-doxylink
pyproject-hooks==1.1.0
    # via
    #   build
    #   pip-tools
pytest==8.3.2
    # via rocm-docs-core (pyproject.toml)
python-dateutil==2.9.0.post0
    # via sphinxcontrib-doxylink
pyyaml==6.0.1
    # via
    #   commitizen
    #   myst-parser
    #   pre-commit
    #   rocm-docs-core (pyproject.toml)
    #   sphinx-external-toc
questionary==2.0.1
    # via commitizen
requests==2.32.3
    # via
    #   pygithub
    #   sphinx
ruff==0.5.5
    # via rocm-docs-core (pyproject.toml)
six==1.16.0
    # via python-dateutil
smmap==5.0.1
    # via gitdb
snowballstemmer==2.2.0
    # via sphinx
soupsieve==2.5
    # via beautifulsoup4
sphinx==7.4.7
    # via
    #   breathe
    #   myst-parser
    #   pydata-sphinx-theme
    #   rocm-docs-core (pyproject.toml)
    #   sphinx-book-theme
    #   sphinx-copybutton
    #   sphinx-design
    #   sphinx-external-toc
    #   sphinx-notfound-page
    #   sphinxcontrib-doxylink
sphinx-book-theme==1.1.3
    # via rocm-docs-core (pyproject.toml)
sphinx-copybutton==0.5.2
    # via rocm-docs-core (pyproject.toml)
sphinx-design==0.6.0
    # via rocm-docs-core (pyproject.toml)
sphinx-external-toc==1.0.1
    # via rocm-docs-core (pyproject.toml)
sphinx-notfound-page==1.0.4
    # via rocm-docs-core (pyproject.toml)
sphinxcontrib-applehelp==2.0.0
    # via sphinx
sphinxcontrib-devhelp==2.0.0
    # via sphinx
sphinxcontrib-doxylink==1.12.3
    # via rocm-docs-core (pyproject.toml)
sphinxcontrib-htmlhelp==2.1.0
    # via sphinx
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-qthelp==2.0.0
    # via sphinx
sphinxcontrib-serializinghtml==2.0.0
    # via sphinx
termcolor==2.4.0
    # via commitizen
tomli==2.0.1
    # via
    #   black
    #   build
    #   mypy
    #   pip-tools
    #   pytest
    #   sphinx
tomlkit==0.13.0
    # via commitizen
tqdm==4.66.4
    # via mpire
typing-extensions==4.12.2
    # via
    #   black
    #   mypy
    #   pydata-sphinx-theme
    #   pygithub
urllib3==2.2.2
    # via
    #   pygithub
    #   requests
virtualenv==20.26.3
    # via pre-commit
wcwidth==0.2.13
    # via prompt-toolkit
wheel==0.43.0
    # via pip-tools
wrapt==1.16.0
    # via deprecated
zipp==3.19.2
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
